[document]添加RV类型-实体-布局对应文档；添加数据源接口示例json

+++ app/src/main/java/com/skyd/imomoe/view/component/player/AnimeVideoPositionMemoryStore.kt
@@ -20,5 +20,5 @@ object AnimeVideoPositionMemoryStore : AnimeVideoPlayer.PlayPositionMemoryDataSt
 
     override fun positionFormat(position: Long): String =
         if (position < 0L) appContext.getString(R.string.episode_play_completed)
-        else CommonUtil.stringForTime(position.toInt())
+        else CommonUtil.stringForTime(position)
 }
